# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HBECode API",
    "version": "0.6.0",
    "description": "Cookie-session API. Unsafe methods require the `x-csrf-token` header (from GET /auth/csrf) and an allowed Origin. Executor endpoints (/internal/executor/*) are bearer-token only and are intentionally not documented here. Realtime pushes: WebSocket at /api/v1/ws (session cookie + allowed Origin); send {\"op\":\"sub\",\"channel\":\"attempt\",\"attemptId\",\"token\"} or {\"op\":\"sub\",\"channel\":\"monitor\",\"testId\"}."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/tests/{id}": {
      "get": {
        "operationId": "getTestsId",
        "tags": [
          "tests"
        ],
        "summary": "Test details",
        "description": "Requires permission `test:proctor`.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "putTestsId",
        "tags": [
          "tests"
        ],
        "summary": "Edit a draft test",
        "description": "Requires permission `test:manage`.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 200
                  },
                  "description": {
                    "default": "",
                    "type": "string",
                    "maxLength": 5000
                  },
                  "startsAt": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                  },
                  "endsAt": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
                  },
                  "durationMin": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1440
                  },
                  "questions": {
                    "minItems": 1,
                    "maxItems": 50,
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "questionId": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "points": {
                          "default": 100,
                          "type": "number",
                          "exclusiveMinimum": 0,
                          "maximum": 1000
                        }
                      },
                      "required": [
                        "questionId"
                      ]
                    }
                  },
                  "settings": {
                    "default": {
                      "requireFullscreen": true,
                      "blockClipboard": true,
                      "webcam": "off",
                      "violations": {
                        "warnAt": 3,
                        "finalWarnAt": 5,
                        "autoSubmitAt": 7
                      },
                      "showResults": true
                    },
                    "type": "object",
                    "properties": {
                      "requireFullscreen": {
                        "default": true,
                        "type": "boolean"
                      },
                      "blockClipboard": {
                        "default": true,
                        "type": "boolean"
                      },
                      "webcam": {
                        "default": "off",
                        "type": "string",
                        "enum": [
                          "off",
                          "flagged"
                        ]
                      },
                      "violations": {
                        "default": {
                          "warnAt": 3,
                          "finalWarnAt": 5,
                          "autoSubmitAt": 7
                        },
                        "type": "object",
                        "properties": {
                          "warnAt": {
                            "default": 3,
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100
                          },
                          "finalWarnAt": {
                            "default": 5,
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100
                          },
                          "autoSubmitAt": {
                            "default": 7,
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 100
                          }
                        }
                      },
                      "showResults": {
                        "default": true,
                        "type": "boolean"
                      }
                    }
                  }
                },
                "required": [
                  "title",
                  "startsAt",
                  "endsAt",
                  "durationMin",
                  "questions"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteTestsId",
        "tags": [
          "tests"
        ],
        "summary": "Delete a draft test",
        "description": "Requires permission `test:manage`.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/overview": {
      "get": {
        "operationId": "getReportsOverview",
        "tags": [
          "reports"
        ],
        "summary": "Institution dashboard: people, content, 30-day activity, recent tests, flagged questions",
        "description": "Requires role client_admin | teacher | associate.",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/tests/{id}": {
      "get": {
        "operationId": "getReportsTestsId",
        "tags": [
          "reports"
        ],
        "summary": "Test report: summary, score distribution, per-question and per-student results, similarity summary",
        "description": "Requires role client_admin | teacher | associate.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/tests/{id}/export": {
      "get": {
        "operationId": "getReportsTestsIdExport",
        "tags": [
          "reports"
        ],
        "summary": "Test report as CSV or Excel (audited)",
        "description": "Requires role client_admin | teacher.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "enum": [
                "csv",
                "xlsx"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/questions/{id}": {
      "get": {
        "operationId": "getReportsQuestionsId",
        "tags": [
          "reports"
        ],
        "summary": "Question report: acceptance, solve rate, by language, verdicts, 30-day activity, tests",
        "description": "Requires role client_admin | teacher | associate.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/batches/{id}": {
      "get": {
        "operationId": "getReportsBatchesId",
        "tags": [
          "reports"
        ],
        "summary": "Batch report: students × tests grid",
        "description": "Requires role client_admin | teacher | associate.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/batches/{id}/export": {
      "get": {
        "operationId": "getReportsBatchesIdExport",
        "tags": [
          "reports"
        ],
        "summary": "Batch report as CSV or Excel (audited)",
        "description": "Requires role client_admin | teacher.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$schema": "https://json-schema.org/draft/2020-12/schema",
              "type": "string",
              "enum": [
                "csv",
                "xlsx"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/students/{id}": {
      "get": {
        "operationId": "getReportsStudentsId",
        "tags": [
          "reports"
        ],
        "summary": "Student progress (staff of the institution, or the student themselves)",
        "description": "Requires permission `report:view`.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/me": {
      "get": {
        "operationId": "getReportsMe",
        "tags": [
          "reports"
        ],
        "summary": "My progress (scores only where the test releases results)",
        "description": "Requires permission `report:view`.",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/platform": {
      "get": {
        "operationId": "getReportsPlatform",
        "tags": [
          "reports"
        ],
        "summary": "Platform health: institutions, daily totals, queue depth, executors",
        "description": "Requires role super_admin.",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/rebuild": {
      "post": {
        "operationId": "postReportsRebuild",
        "tags": [
          "reports"
        ],
        "summary": "Recompute all report rollups from raw data (audited)",
        "description": "Requires role super_admin.",
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tests/{id}/plagiarism": {
      "post": {
        "operationId": "postTestsIdPlagiarism",
        "tags": [
          "reports"
        ],
        "summary": "Start a code-similarity check for a test (runs in the background; audited). scope=institution also compares with the same questions in the institution's other tests",
        "description": "Requires permission `test:manage`.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "scope": {
                    "default": "test",
                    "type": "string",
                    "enum": [
                      "test",
                      "institution"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getTestsIdPlagiarism",
        "tags": [
          "reports"
        ],
        "summary": "Latest similarity check and its flagged pairs",
        "description": "Requires permission `test:proctor`.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plagiarism/{runId}/pairs/{questionId}/{subA}/{subB}": {
      "get": {
        "operationId": "getPlagiarismRunIdPairsQuestionIdSubASubB",
        "tags": [
          "reports"
        ],
        "summary": "Both submissions of a flagged pair with matching line ranges",
        "description": "Requires permission `test:proctor`.",
        "security": [
          {
            "session": []
          }
        ],
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subA",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subB",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hb_at"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}